// src/types/openapi_MS.ts
export interface components {
    schemas: {
// ...
        ReqTimetableTemplateItemDTO: {
            /** Format: uuid */
            lessonTypeId: string;
            /** @enum {string} */
            dayOfWeek: "MONDAY" | "TUESDAY" | "WEDNESDAY" | "THURSDAY" | "FRIDAY" | "SATURDAY" | "SUNDAY";
            startTime: string;
            /** Format: int32 */
            sortOrder: number;
        };
// ...
        ReqEmployeeRATemplateItemDTO: {
            /** Format: uuid */
            lessonTypeId: string;
            /** Format: uuid */
            userUuid: string;
            /** @enum {string} */
            dayOfWeek: "MONDAY" | "TUESDAY" | "WEDNESDAY" | "THURSDAY" | "FRIDAY" | "SATURDAY" | "SUNDAY";
            startTime: string;
            /** Format: int32 */
            sortOrder: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}